+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ## Fixed
+
 - CopySlice wraparound issue especially during CopyCallBytesToMemory
+- EVM.Solidity.toCode to include contractName in error string
 
 ## Changed